pyproject.toml
[tool.poetry]
name = "cwubs-shared"
version = "0.0.0" # placeholder
description = "Shared utilities for cwubs services"
authors = ["Duncan Scuffell <duncan.scuffell@btinternet.com>"]
readme = "README.md"
packages = [{ include = "shared" }]

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
sqlalchemy = "^2.0.46"
fastapi = "^0.128.5"
pyjwt = "^2.11.0"
loguru = "^0.7.3"

[tool.poetry-dynamic-versioning]
enable = true

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning>=1.0.0,<2.0.0"]
build-backend = "poetry_dynamic_versioning.backend"

[tool.poetry.requires-plugins]
poetry-dynamic-versioning = { version = ">=1.0.0,<2.0.0", extras = ["plugin"] }
